Make group IDs 64 bit

+++ pkg/mongo/player.go
@@ -37,7 +37,7 @@ type Player struct {
 	Friends          string    `bson:"friends"`          // []ProfileFriend
 	GamesRecent      string    `bson:"games_recent"`     // []ProfileRecentGame
 	GameStats        string    `bson:"game_stats"`       // PlayerAppStatsTemplate
-	Groups           []int     `bson:"groups"`           // []int
+	Groups           []int64   `bson:"groups"`           // []int
 	LastLogOff       time.Time `bson:"time_logged_off"`  //
 	NumberOfGameBans int       `bson:"bans_game"`        //
 	NumberOfVACBans  int       `bson:"bans_cav"`         //
